+++ projects/demo-auth/src/app/components/global-alert/global-alert.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { IAlert } from './alert.service';
 
 import { ClarityIcons, infoCircleIcon, windowCloseIcon} from '@cds/core/icon';
-import { NgIf, NgClass, NgFor } from '@angular/common';
+import { NgClass } from '@angular/common';
 import { ClrIconModule, ClrAlertModule } from '@clr/angular';
 ClarityIcons.addIcons(...[infoCircleIcon, windowCloseIcon]);
 
@@ -11,7 +11,7 @@ ClarityIcons.addIcons(...[infoCircleIcon, windowCloseIcon]);
     templateUrl: './global-alert.component.html',
     styleUrls: ['./global-alert.component.scss'],
     standalone: true,
-    imports: [NgIf, NgClass, ClrIconModule, ClrAlertModule, NgFor]
+    imports: [NgClass, ClrIconModule, ClrAlertModule]
 })
 export class GlobalAlertComponent {
   @Input() alert!: null | IAlert;