+++ projects/demo-maps/src/app/app.component.ts
@@ -9,7 +9,7 @@ import { appVersion } from '../environments/version';
 import { Subscription } from 'rxjs';
 import { filter } from 'rxjs/operators';
 import { ClrMainContainerModule, ClrNavigationModule } from '@clr/angular';
-import { NgIf, NgFor } from '@angular/common';
+
 import { GlobalAlertComponent } from './components/global-alert/global-alert.component';
 import { GlobalProgressComponent } from './components/global-progress/global-progress.component';
 import { HeaderComponent } from './components/header/header.component';
@@ -24,7 +24,7 @@ interface IUi {
     templateUrl: './app.component.html',
     styleUrls: ['./app.component.scss'],
     standalone: true,
-    imports: [ClrMainContainerModule, ClrNavigationModule, NgIf, GlobalAlertComponent, GlobalProgressComponent, HeaderComponent, RouterLink, RouterLinkActive, NgFor, RouterOutlet]
+    imports: [ClrMainContainerModule, ClrNavigationModule, GlobalAlertComponent, GlobalProgressComponent, HeaderComponent, RouterLink, RouterLinkActive, RouterOutlet]
 })
 export class AppComponent implements OnInit, OnDestroy {
   title = '';