+++ projects/demo-maps/src/app/components/example-group-action/example-group-action.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { LayerGroup, LayersService, ILayerGroupOptions } from '@dlr-eoc/services-layers';
-import { NgFor } from '@angular/common';
+
 
 @Component({
     selector: 'app-example-group-action',
     templateUrl: './example-group-action.component.html',
     styleUrls: ['./example-group-action.component.scss'],
     standalone: true,
-    imports: [NgFor]
+    imports: []
 })
 export class ExampleGroupActionComponent implements OnInit {
   @Input() group: LayerGroup;