+++ projects/demo-maps/src/app/components/global-progress/global-progress.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit, Input, EventEmitter, Output } from '@angular/core';
 import { IProgress } from './progress.service';
-import { NgIf, NgClass } from '@angular/common';
+import { NgClass } from '@angular/common';
 
 @Component({
     selector: 'ukis-global-progress',
     templateUrl: './global-progress.component.html',
     styleUrls: ['./global-progress.component.scss'],
     standalone: true,
-    imports: [NgIf, NgClass]
+    imports: [NgClass]
 })
 export class GlobalProgressComponent implements OnInit {
   @Input() progress!: null | IProgress;