+++ projects/demo-maps/src/app/components/header/header.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit, Input } from '@angular/core';
 import { ClarityIcons} from '@cds/core/icon';
 import { ukisIcon, dlrIcon } from '../icons/ukis';
 import { ClrNavigationModule, ClrIconModule, ClrStandaloneCdkTrapFocus } from '@clr/angular';
-import { NgClass, NgIf } from '@angular/common';
+import { NgClass } from '@angular/common';
 import { RouterLink, RouterLinkActive } from '@angular/router';
 ClarityIcons.addIcons(ukisIcon, dlrIcon);
 
@@ -12,7 +12,7 @@ ClarityIcons.addIcons(ukisIcon, dlrIcon);
     templateUrl: './header.component.html',
     styleUrls: ['./header.component.scss'],
     standalone: true,
-    imports: [ClrNavigationModule, NgClass, ClrIconModule, RouterLink, RouterLinkActive, NgIf, ClrStandaloneCdkTrapFocus]
+    imports: [ClrNavigationModule, NgClass, ClrIconModule, RouterLink, RouterLinkActive, ClrStandaloneCdkTrapFocus]
 })
 export class HeaderComponent implements OnInit {
   @Input('ukis-title') title: string = '';