+++ projects/demo-maps/src/app/components/popup2/popup2.component.ts
@@ -1,12 +1,12 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { NgFor, KeyValuePipe } from '@angular/common';
+import { KeyValuePipe } from '@angular/common';
 
 @Component({
     selector: 'app-popup2',
     templateUrl: './popup2.component.html',
     styleUrls: ['./popup2.component.scss'],
     standalone: true,
-    imports: [NgFor, KeyValuePipe]
+    imports: [KeyValuePipe]
 })
 export class Popup2Component implements OnInit {