+++ projects/demo-maps/src/app/components/table-popup/table-popup.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { ClrDatagridModule, ClrStopEscapePropagationDirective, ClrPopoverHostDirective } from '@clr/angular';
-import { NgFor, KeyValuePipe } from '@angular/common';
+import { KeyValuePipe } from '@angular/common';
 
 @Component({
     selector: 'app-table-popup',
     templateUrl: './table-popup.component.html',
     styleUrls: ['./table-popup.component.scss'],
     standalone: true,
-    imports: [ClrDatagridModule, NgFor, ClrStopEscapePropagationDirective, ClrPopoverHostDirective, KeyValuePipe]
+    imports: [ClrDatagridModule, ClrStopEscapePropagationDirective, ClrPopoverHostDirective, KeyValuePipe]
 })
 export class TablePopupComponent implements OnInit {