+++ projects/demo-maps/src/app/route-components/route-licenses/route-licenses.component.ts
@@ -3,7 +3,7 @@ import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 import { map } from 'rxjs/operators';
 import { ClrStopEscapePropagationDirective, ClrPopoverHostDirective, ClrSignpostModule, ClrConditionalModule, ClrDatagridModule } from '@clr/angular';
-import { NgIf, NgFor, AsyncPipe } from '@angular/common';
+import { AsyncPipe } from '@angular/common';
 
 /**
  * This component expects a assets/licenses.json file to be present in the project bundle.
@@ -54,7 +54,7 @@ interface License {
     templateUrl: './route-licenses.component.html',
     styleUrls: ['./route-licenses.component.scss'],
     standalone: true,
-    imports: [ClrStopEscapePropagationDirective, ClrPopoverHostDirective, ClrSignpostModule, ClrConditionalModule, ClrDatagridModule, NgIf, NgFor, AsyncPipe]
+    imports: [ClrStopEscapePropagationDirective, ClrPopoverHostDirective, ClrSignpostModule, ClrConditionalModule, ClrDatagridModule, AsyncPipe]
 })
 export class RouteLicensesComponent implements OnInit {
   @HostBinding('class') class = 'content-container';