+++ projects/layer-control/src/lib/layer-control/layer-control.component.ts
@@ -6,7 +6,7 @@ import { LayerGroup, Layer } from '@dlr-eoc/services-layers';
 import { Subscription } from 'rxjs';
 
 import { CdkDragDrop, CdkDropList, CdkDrag, CdkDragHandle } from '@angular/cdk/drag-drop';
-import { NgFor, NgIf } from '@angular/common';
+
 import { LayerentryGroupComponent } from '../layerentry-group/layerentry-group.component';
 import { LayerentryComponent } from '../layerentry/layerentry.component';
 import { ItemsFilterPipe } from '../utils/obj-type.pipe';
@@ -17,7 +17,7 @@ import { ReversePipe } from '../utils/array-reverse.pipe';
     templateUrl: './layer-control.component.html',
     styleUrls: ['./layer-control.component.scss'],
     standalone: true,
-    imports: [CdkDropList, NgFor, CdkDrag, NgIf, LayerentryGroupComponent, CdkDragHandle, LayerentryComponent, ItemsFilterPipe, ReversePipe]
+    imports: [CdkDropList, CdkDrag, LayerentryGroupComponent, CdkDragHandle, LayerentryComponent, ItemsFilterPipe, ReversePipe]
 })
 export class LayerControlComponent implements OnInit, OnDestroy {
   @Input('layersSvc') layersSvc: LayersService;