+++ projects/map-tools/src/lib/mouse-position/mouse-position.component.ts
@@ -3,7 +3,7 @@ import { MapOlService } from '@dlr-eoc/map-ol';
 import { transform as olTransform, get as olGetProjection, Projection as olProjection } from 'ol/proj';
 import { Subscription } from 'rxjs';
 import { ClrSelectModule, ClrCommonFormsModule, ClrInputModule } from '@clr/angular';
-import { NgFor } from '@angular/common';
+
 import { FormsModule } from '@angular/forms';
 
 interface ISelectProjection {
@@ -16,7 +16,7 @@ interface ISelectProjection {
     templateUrl: './mouse-position.component.html',
     styleUrls: ['./mouse-position.component.scss'],
     standalone: true,
-    imports: [ClrSelectModule, ClrCommonFormsModule, NgFor, FormsModule, ClrInputModule]
+    imports: [ClrSelectModule, ClrCommonFormsModule, FormsModule, ClrInputModule]
 })
 export class MousePositionComponent implements OnInit, OnDestroy {