+++ projects/services-layers/src/lib/types/Layers.ts
@@ -6,7 +6,7 @@ export interface IAnyObject {
 
 /**
  * When using dynamic components for layers you should keep in mind to handle the state (with a service or object binding) if needed,
- * because `legendIng` or `action` are placed inside `ngIf` directives in the template and are destroyed when they are hidden.
+ * because `legendIng` or `action` are placed inside `@if` directives in the template and are destroyed when they are hidden.
  */
 interface IDynamicComponent {
   component: Type<any>;