+++ projects/user-info/src/lib/user-details/user-details.component.ts
@@ -1,14 +1,14 @@
 import { Component, OnDestroy } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { UserService } from '../user.service';
-import { NgIf } from '@angular/common';
+
 
 @Component({
     selector: 'ukis-user-details',
     templateUrl: './user-details.component.html',
     styleUrls: ['./user-details.component.scss'],
     standalone: true,
-    imports: [NgIf]
+    imports: []
 })
 export class UserDetailsComponent implements OnDestroy {